added "create_markdown"
A new routine, "create_markdown.py" was added that takes a proof written as a LaTeX file and transfers it into a Markdown document (almost) ready for submission to the StatProofBook. This considerably speeds up content preparation - only need to add title, theorem, sources and proof ID at the end.