added "create_def"
A new routine, "create_def.py" was added that takes a definition written as a LaTeX file and transfers it into a Markdown document (almost) ready for submission to the StatProofBook. This considerably speeds up content preparation - only need to add title, theorem, sources and definition ID at the end.